backoffice/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
appnope==0.1.0            # via ipython
asgiref==3.2.10           # via -r requirements/base.txt, django
attrs==19.3.0             # via pytest
backcall==0.2.0           # via ipython
certifi==2020.6.20        # via -r requirements/base.txt, requests, sentry-sdk
chardet==3.0.4            # via -r requirements/base.txt, requests
coverage==5.2             # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
django-annoying==0.10.6   # via -r requirements/base.txt, django-viewflow
django-appconf==1.0.4     # via -r requirements/base.txt, django-compressor
django-compressor==2.4    # via -r requirements/base.txt
django-debug-toolbar==3.1.1  # via -r requirements/dev.in
django-environ==0.4.5     # via -r requirements/base.txt
django-extensions==3.0.9  # via -r requirements/dev.in
django-filter==2.3.0      # via -r requirements/base.txt, django-viewflow
django-jsonstore==0.4.1   # via -r requirements/base.txt, django-viewflow
django-material==1.7.1    # via -r requirements/base.txt
django-phonenumber-field==5.0.0  # via -r requirements/base.txt
django-sass-processor==0.8.2  # via -r requirements/base.txt
git+https://github.com/uktrade/django-staff-sso-client.git@master#egg=django-staff-sso-client  # via -r requirements/base.txt
django-viewflow==1.6.1    # via -r requirements/base.txt
django==3.1.1             # via -r requirements/base.txt, django-annoying, django-appconf, django-debug-toolbar, django-filter, django-jsonstore, django-phonenumber-field, django-staff-sso-client, django-viewflow, djangorestframework
djangorestframework==3.11.1  # via -r requirements/base.txt
docopt==0.6.2             # via -r requirements/base.txt, notifications-python-client
factory-boy==3.0.1        # via -r requirements/dev.in
faker==4.1.1              # via factory-boy
flake8==3.8.3             # via -r requirements/dev.in
future==0.18.2            # via -r requirements/base.txt, notifications-python-client
gunicorn==20.0.4          # via -r requirements/base.txt
httpretty==1.0.2          # via -r requirements/dev.in
idna==2.10                # via -r requirements/base.txt, requests
iniconfig==1.0.1          # via pytest
ipdb==0.13.3              # via -r requirements/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via ipdb
jedi==0.17.1              # via ipython
libsass==0.20.1           # via -r requirements/base.txt
mccabe==0.6.1             # via flake8
monotonic==1.5            # via -r requirements/base.txt, notifications-python-client
more-itertools==8.4.0     # via pytest
notifications-python-client==5.7.0  # via -r requirements/base.txt
oauthlib==3.1.0           # via -r requirements/base.txt, requests-oauthlib
packaging==20.4           # via pytest, pytest-sugar
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
phonenumbers==8.12.10     # via -r requirements/base.txt
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
psycopg2==2.8.6           # via -r requirements/base.txt
ptyprocess==0.6.0         # via pexpect
py==1.8.2                 # via pytest
pycodestyle==2.6.0        # via flake8
pydotplus==2.0.2          # via -r requirements/dev.in
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyjwt==1.7.1              # via -r requirements/base.txt, notifications-python-client
pyparsing==2.4.7          # via packaging, pydotplus
pytest-cov==2.10.1        # via -r requirements/dev.in
pytest-django==3.10.0     # via -r requirements/dev.in
pytest-dotenv==0.5.2      # via -r requirements/dev.in
pytest-sugar==0.9.4       # via -r requirements/dev.in
pytest==6.0.2             # via -r requirements/dev.in, pytest-cov, pytest-django, pytest-dotenv, pytest-sugar
python-dateutil==2.8.1    # via faker
python-dotenv==0.13.0     # via pytest-dotenv
pytz==2020.1              # via -r requirements/base.txt, django
rcssmin==1.0.6            # via -r requirements/base.txt, django-compressor
requests-oauthlib==1.3.0  # via -r requirements/base.txt, django-staff-sso-client
requests==2.24.0          # via -r requirements/base.txt, notifications-python-client, requests-oauthlib
rjsmin==1.1.0             # via -r requirements/base.txt, django-compressor
sentry-sdk==0.17.8        # via -r requirements/base.txt
six==1.15.0               # via -r requirements/base.txt, django-annoying, django-compressor, django-jsonstore, django-material, django-viewflow, libsass, packaging, python-dateutil, traitlets
sqlparse==0.3.1           # via -r requirements/base.txt, django, django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
testfixtures==6.14.2      # via -r requirements/dev.in
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest
traitlets==4.3.3          # via ipython
urllib3==1.25.9           # via -r requirements/base.txt, requests, sentry-sdk
wcwidth==0.2.4            # via prompt-toolkit
whitenoise==5.2.0         # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
